+++ src/genome/aligner/__init__.py
@@ -1,7 +1,8 @@
 """Aligner abstractions and the :class:`AlignerMixin` for :class:`~genome.genome.Genome`."""
 
 from genome.aligner.aligner import Aligner
+from genome.aligner.chromap import Chromap
 from genome.aligner.mixin import AlignerMixin
 from genome.aligner.star import STAR
 
-__all__ = ["STAR", "Aligner", "AlignerMixin"]
+__all__ = ["STAR", "Aligner", "AlignerMixin", "Chromap"]